Rename to_data_type to to_candid_type
