+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_fn_or_constructor_type/azle_fn_type/mod.rs
@@ -5,7 +5,7 @@ mod errors;
 mod function_and_method_type_helper_methods;
 mod get_source_info;
 mod get_source_text;
-mod to_data_type;
+mod to_candid_type;
 
 #[derive(Clone)]
 pub struct AzleFnType<'a> {
