+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_keyword_type/mod.rs
@@ -4,7 +4,7 @@ use swc_ecma_ast::TsKeywordType;
 mod errors;
 mod get_source_info;
 mod get_source_text;
-mod to_data_type;
+mod to_candid_type;
 
 #[derive(Clone)]
 pub struct AzleKeywordType<'a> {
