+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/to_candid_type.rs
@@ -4,7 +4,7 @@ use super::AzleType;
 use crate::ts_ast::traits::GetSourceInfo;
 
 impl AzleType<'_> {
-    pub fn to_data_type(&self) -> CandidType {
+    pub fn to_candid_type(&self) -> CandidType {
         match self {
             AzleType::AzleKeywordType(azle_keyword_type) => azle_keyword_type.to_data_type(),
             AzleType::AzleTypeRef(azle_type_ref) => azle_type_ref.to_candid_type(),
